Refactor Sold class
